# docs/openapi.yml
openapi: 3.0.3
info:
  title: API - Sistema de Citas Médicas
  version: 1.0.0
  description: >
    Contrato central de APIs para el sistema de citas médicas basado en microservicios.
    Las rutas se publican a través del API Gateway.
servers:
  - url: http://localhost:9090
    description: API Gateway local
...
paths:
...
  /ms-cita/api/v1/agendamiento/confirmar:
    post:
      tags: [Citas]
      summary: Confirmar cita médica
      operationId: confirmarCita
      description: >
        Valida el paciente y la disponibilidad del médico, registra la cita
        y genera una notificación de confirmación.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CitaRequest'
            example:
              pacienteId: 1
              colegiaturaMedico: "CMP001"
              turnoCita: "09:00"
              costoConsulta: 100.0
      responses:
        '201':
          description: Cita registrada correctamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CitaResponse'
        '400':
          description: Datos de cita inválidos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Paciente no encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Médico sin disponibilidad
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CitaRequest:
      type: object
      required: [pacienteId, colegiaturaMedico, turnoCita, costoConsulta]
      properties:
        pacienteId:
          type: integer
          format: int32
          example: 1
        colegiaturaMedico:
          type: string
          example: "CMP001"
        turnoCita:
          type: string
          example: "09:00"
        costoConsulta:
          type: number
          format: double
          minimum: 0.01
          example: 100.0
    CitaResponse:
      allOf:
        - type: object
          properties:
            id:
              type: integer
              format: int32
              example: 10
        - $ref: '#/components/schemas/CitaRequest'
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 404
        error:
          type: string
          example: "Not Found"
        message:
          type: string
          example: "Recurso no encontrado"
        path:
          type: string
          example: "/ms-paciente/api/v1/pacientes/99"
security:
  - bearerAuth: []